ai-interviewer/requirements.txt
# AI Interviewer Requirements

# Core dependencies
streamlit==1.32.0
SpeechRecognition==3.10.0
openai==1.12.0
textstat==0.7.3
gtts==2.5.0
pydub==0.25.1
python-dotenv==1.0.0
anthropic==0.8.0
google-generativeai==0.3.1
reportlab==4.0.4
numpy==1.24.3
matplotlib==3.7.2
pytest==7.4.0
pyaudio==0.2.14

# Note: PyAudio requires PortAudio as a system dependency
# Install PortAudio before installing PyAudio:
# - macOS: brew install portaudio
# - Ubuntu/Debian: sudo apt-get install portaudio19-dev
# - Windows: pip install pipwin && pipwin install pyaudio